backend/todo_back/requirements.txt
Django==4.2.6
django-extensions==3.2.3
djangorestframework==3.14.0
ipdb==0.13.13
ipython==8.16.1
Werkzeug==3.0.0